bugfix: Set item comments when parsing geeklists
When accessing a geeklist with comments enabled, the parsing of the XML structure was broken. It would always add the comments on the main Geeklist itself to every item in the list. This had the worse side-effect of losing all comments on the items themselves.

Fix the logic bug, so that every item's own comments are stored in its comments instance.